move all tests that are transaction specific from client_tests
